+++ Player/Tech/HLS/HLSNative.swift
@@ -255,13 +255,21 @@ public final class HLSNative<Context: MediaContext>: PlaybackTech {
     
     /// Enabling this function will cause HLSNative to continuously dispatch all error events encountered, including recoverable errors not resulting in playback to stop, to the associated `MediaSource`s analytics providers.
     ///
-    /// - Important: Not recomended for use in production, will generate high volume traffic. For debug purposes only.
+    /// - Important: DEPRECATED in 2.0.86 in favour of DEBUG only `LogLevel`.
+    @available(*, message: "DEPRECATED in favour of DEBUG only LogLevel")
     public var continuouslyDispatchErrorLogEvents: Bool = false
+
+    /// Internal log level
+    internal var logLevel: LogLevel = .none
     
-    /// Enabling this function will cause HLSNative to continuously dispatch all access log events encountered.
-    ///
-    /// - Important: Not recomended for use in production, will generate high volume traffic. For debug purposes only.
-    public var continuouslyDispatchAccessLogEvents: Bool = false
+    /// LogLevel option
+    internal enum LogLevel {
+        /// No logging
+        case none
+        
+        /// HLSNative will continuously print all error and access log events encountered to the console, including recoverable errors not resulting in playback failure.
+        case debug
+    }
 }
 
 // MARK: - Load Source
@@ -730,7 +738,7 @@ extension HLSNative {
 
 extension HLSNative {
     fileprivate func handleNewAccessLogEntry(mediaAsset: MediaAsset<Context.Source>) {
-        if continuouslyDispatchAccessLogEvents {
+        if logLevel == .debug {
             // IMPORTANT: Do not use in a production environment
             let playerItem = mediaAsset.playerItem
             mediaAsset.itemObserver.subscribe(notification: .AVPlayerItemNewAccessLogEntry, for: playerItem) { [weak self] notification in
@@ -739,29 +747,25 @@ extension HLSNative {
                     var json: [String: Any] = [
                         "Message": "PLAYER_ITEM_ACCESS_LOG_ENTRY"
                     ]
-                    var info = "StartupTime: \(Int64(event.startupTime)) \n"
-                    info += "NumberOfStalls: \(event.numberOfStalls) \n"
-                    info += "NumberOfDroppedVideoFrames: \(event.numberOfDroppedVideoFrames) \n"
-                    info += "DownloadOverdue: \(event.downloadOverdue) \n"
-                    info += "NumberOfServerAddressChanges: \(event.numberOfServerAddressChanges) \n"
-                    info += "MediaRequestsWWAN: \(event.mediaRequestsWWAN) \n"
-                    info += "TransferDuration: \(event.transferDuration) \n"
-                    info += "MediaRequests: \(event.numberOfMediaRequests) \n"
+                    var info = "StartupTime: \(Int64(event.startupTime)) \n "
+                    info += "NumberOfStalls: \(event.numberOfStalls) \n "
+                    info += "NumberOfDroppedVideoFrames: \(event.numberOfDroppedVideoFrames) \n "
+                    info += "DownloadOverdue: \(event.downloadOverdue) \n "
+                    info += "NumberOfServerAddressChanges: \(event.numberOfServerAddressChanges) \n "
+                    info += "MediaRequestsWWAN: \(event.mediaRequestsWWAN) \n "
+                    info += "TransferDuration: \(event.transferDuration) \n "
+                    info += "MediaRequests: \(event.numberOfMediaRequests) \n "
                     
                     if let uri = event.uri {
                         info += "URI: \(uri) \n"
                     }
                     
                     if let serverAddress = event.serverAddress {
-                        info += "ServerAddress: \(serverAddress) \n"
+                        info += "ServerAddress: \(serverAddress) \n "
                     }
                     json["Info"] = info
                     
-                    mediaAsset.source.analyticsConnector.providers.forEach {
-                        if let traceProvider = $0 as? TraceProvider {
-                            traceProvider.onTrace(tech: self, source: mediaAsset.source, data: json)
-                        }
-                    }
+                    print(json)
                 }
             }
         }
@@ -791,10 +795,9 @@ extension HLSNative {
                         self.stop()
                     }
                 }
-                else {
-                    if let mediaAsset = self.currentAsset, self.continuouslyDispatchErrorLogEvents {
-                        mediaAsset.source.analyticsConnector.onTrace(tech: self, source: mediaAsset.source, data: event.traceProviderData)
-                    }
+                
+                if self.logLevel == .debug {
+                    print(event.traceProviderData)
                 }
             }
         }