+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
 
+* `2.0.86` Release - [2.0.86](#2086)
 * `2.0.85` Release - [2.0.85](#2085)
 * `2.0.81` Release - [2.0.81](#2081)
 * `2.0.80` Release - [2.0.80](#2080)
@@ -11,10 +12,10 @@
 * `0.2.x` Releases - [0.2.0](#020)
 * `0.1.x` Releases - [0.1.0](#010) | [0.1.1](#011) | [0.1.2](#012) | [0.1.3](#013) | [0.1.4](#014) | [0.1.5](#015)
 
-## Unreleased
+## 2.0.86
 
-#### Features
-* Added feature to continuously dispatch access and error log entries to analytics. Note: For **debug** puposes only.
+#### Changes
+* Added a `DEBUG` feature to continuously print access and error log entries.
 
 ## 2.0.85